Use auto where possible; Change add* measures' signatures to return opt<int>

+++ src/cpp_src/pb_methods_and_measures/GreedyOverCost.cpp
@@ -77,8 +77,8 @@ std::optional<long long> cost_reduction_for_greedy_over_cost(const Election &ele
     return max_price_to_be_chosen;
 }
 
-std::optional<long long> optimist_add_for_greedy_over_cost(const Election &election, int p,
-                                                           const ProjectComparator &tie_breaking) {
+std::optional<int> optimist_add_for_greedy_over_cost(const Election &election, int p,
+                                                     const ProjectComparator &tie_breaking) {
     auto total_budget = election.budget();
     auto num_voters = election.numVoters();
     auto projects = election.projects();
@@ -121,13 +121,13 @@ std::optional<long long> optimist_add_for_greedy_over_cost(const Election &elect
     return {}; // LCOV_EXCL_LINE (every project should be feasible)
 }
 
-std::optional<long long> pessimist_add_for_greedy_over_cost(const Election &election, int p,
-                                                            const ProjectComparator &tie_breaking) {
+std::optional<int> pessimist_add_for_greedy_over_cost(const Election &election, int p,
+                                                      const ProjectComparator &tie_breaking) {
     return optimist_add_for_greedy_over_cost(election, p, tie_breaking);
 }
 
-std::optional<long long> singleton_add_for_greedy_over_cost(const Election &election, int p,
-                                                            const ProjectComparator &tie_breaking) {
+std::optional<int> singleton_add_for_greedy_over_cost(const Election &election, int p,
+                                                      const ProjectComparator &tie_breaking) {
     auto total_budget = election.budget();
     auto projects = election.projects();
     auto pp = projects[p];
